w3c-reports/12092014/gtk/runned-tests/quirks-mode/percentage-height-calculation-expected.txt

    

PASS The percentage height calculation quirk, #test { height:100% }<div id=test></div> 
PASS The percentage height calculation quirk, #test { height:50% }<div id=test></div> 
PASS The percentage height calculation quirk, #test { height:25% }<div id=test></div> 
PASS The percentage height calculation quirk, #test { height:12.5% }<div id=test></div> 
PASS The percentage height calculation quirk, #test { height:100% }<div><div id=test></div></div> 
FAIL The percentage height calculation quirk, <img id=test src="{png}" height=100%> assert_equals: quirks mode expected "184px" but got "185px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:92:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:82:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:81:22)
FAIL The percentage height calculation quirk, <img id=test src="{png}" height=100% border=10> assert_equals: quirks mode expected "184px" but got "185px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:92:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:82:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:81:22)
PASS The percentage height calculation quirk, <table id=test height=100%><tr><td></table> 
PASS The percentage height calculation quirk, #foo { height:100px } #test { height:100% }<div id=foo><div><div id=test></div></div></div> 
PASS The percentage height calculation quirk, #foo { position:absolute } #test { height:100% }<div id=foo><div><div id=test></div></div></div> 
PASS The percentage height calculation quirk, #foo { position:relative } #test { height:100% }<div id=foo><div><div id=test></div></div></div> 
PASS The percentage height calculation quirk, #foo { height:100px } #test { height:100%; position:absolute }<div id=foo><div><div id=test></div></div></div> 
PASS The percentage height calculation quirk, #foo { height:100px } #test { height:100%; position:fixed }<div id=foo><div><div id=test></div></div></div> 
PASS The percentage height calculation quirk, #foo { height:100px } #test { height:100%; position:relative }<div id=foo><div><div id=test></div></div></div> 
PASS The percentage height calculation quirk, #foo { height:100px } #test { height:calc(100% + 1px) }<div id=foo><div id=test></div></div> 
FAIL The percentage height calculation quirk, #foo { height:100px } #test { height:5px; height:calc(100% + 1px) }<div id=foo><div><div id=test></div></div></div> assert_equals: quirks mode expected "0px" but got "101px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:92:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:82:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:81:22)
PASS The percentage height calculation quirk, html { display:inline } #test { height:100% }<div id=test></div> 
FAIL The percentage height calculation quirk, html { margin:10px } body { display:inline } #test { height:100% }<div id=test></div> assert_equals: quirks mode expected "200px" but got "180px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:92:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:82:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:81:22)
PASS The percentage height calculation quirk, body { margin:0 } #test { height:100% }<div id=test></div> 
PASS The percentage height calculation quirk, body { margin:0; padding:10px } #test { height:100% }<div id=test></div> 
PASS The percentage height calculation quirk, body { margin:0; border:10px solid } #test { height:100% }<div id=test></div> 
PASS The percentage height calculation quirk, html { margin:10px } #test { height:100% }<div id=test></div> 
PASS The percentage height calculation quirk, html { padding:10px } #test { height:100% }<div id=test></div> 
PASS The percentage height calculation quirk, html { border:10px solid } #test { height:100% }<div id=test></div> 
PASS The percentage height calculation quirk, html { position:absolute } #test { height:100% }<div id=test></div> 
PASS The percentage height calculation quirk, body { position:absolute } #test { height:100% }<div id=test></div> 
PASS The percentage height calculation quirk, html { height:100%; margin:10px } #test { height:100% }<div id=test></div> 
FAIL The percentage height calculation quirk, html { height:100%; padding:10px } #test { height:100% }<div id=test></div> assert_equals: quirks mode expected "184px" but got "169px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:92:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:82:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:81:22)
FAIL The percentage height calculation quirk, html { height:100%; border:10px solid } #test { height:100% }<div id=test></div> assert_equals: quirks mode expected "184px" but got "169px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:92:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:82:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:81:22)
PASS The percentage height calculation quirk, body { height:100%; margin:10px } #test { height:100% }<div id=test></div> 
FAIL The percentage height calculation quirk, body { height:100%; padding:10px } #test { height:100% }<div id=test></div> assert_equals: quirks mode expected "200px" but got "185px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:92:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:82:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:81:22)
FAIL The percentage height calculation quirk, body { height:100%; border:10px solid } #test { height:100% }<div id=test></div> assert_equals: quirks mode expected "200px" but got "185px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:92:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:82:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:81:22)
FAIL The percentage height calculation quirk, html { position:absolute; height:100%; margin:10px } #test { height:100% }<div id=test></div> assert_equals: quirks mode expected "184px" but got "169px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:92:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:82:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:81:22)
FAIL The percentage height calculation quirk, html { position:absolute; height:100%; padding:10px } #test { height:100% }<div id=test></div> assert_equals: quirks mode expected "184px" but got "169px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:92:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:82:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:81:22)
FAIL The percentage height calculation quirk, html { position:absolute; height:100%; border:10px solid } #test { height:100% }<div id=test></div> assert_equals: quirks mode expected "184px" but got "169px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:92:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:82:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:81:22)
PASS The percentage height calculation quirk, body { margin:99px 0 } #test { height:100% }<div id=test></div> 
PASS The percentage height calculation quirk, body { margin:110px 0 } #test { height:100% }<div id=test></div> 
PASS The percentage height calculation quirk, html, body { border:10px none } #test { height:100% }<div id=test></div> 
PASS The percentage height calculation quirk, html, body { border:10px hidden } #test { height:100% }<div id=test></div> 
PASS The percentage height calculation quirk, <html xmlns="{html}"><head><style>#test { height:100% }</style></head><body><div id="test"/></body></html> 
PASS The percentage height calculation quirk, <html xmlns="{html}"><head><style>#test { height:100% }</style></head><body/><div id="test"/></html> 
FAIL The percentage height calculation quirk, <html xmlns="{html}"><head><style>#test { height:100% }</style></head><span><body><div id="test"/></body></span></html> assert_equals: quirks mode expected "200px" but got "184px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:126:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:119:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:118:26)
FAIL The percentage height calculation quirk, <html xmlns="{html}"><head><style>#test { height:100% }</style></head><body><body><div id="test"/></body></body></html> assert_equals: quirks mode expected "200px" but got "168px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:126:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:119:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:118:26)
FAIL The percentage height calculation quirk, <html><head xmlns="{html}"><style>#test { height:100% }</style></head><body xmlns="{html}"><div id="test"/></body></html> assert_equals: quirks mode expected "200px" but got "184px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:126:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:119:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:118:26)
FAIL The percentage height calculation quirk, <div xmlns="{html}"><head><style>#test { height:100% }</style></head><body><div id="test"/></body></div> assert_equals: quirks mode expected "200px" but got "184px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:126:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:119:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:118:26)
PASS The percentage height calculation quirk, <html xmlns="{html}"><head><style>#test { height:100% }</style></head><body xmlns=""><div xmlns="{html}" id="test"/></body></html> 
FAIL The percentage height calculation quirk, <HTML xmlns="{html}"><head><style>#test { height:100% }</style></head><body><div id="test"/></body></HTML> assert_equals: quirks mode expected "200px" but got "184px"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/quirks-mode/percentage-height-calculation.html:126:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/quirks-mode/percentage-height-calculation.html:119:17
forEach@[native code]
onload@http://localhost:8800/quirks-mode/percentage-height-calculation.html:118:26)
PASS The percentage height calculation quirk, <html xmlns="{html}"><head><style>#test { height:100% }</style></head><BODY><div id="test"/></BODY></html> 
